requirements.txt
asgiref==3.5.2
astroid==2.11.7
backports.zoneinfo==0.2.1
colorama==0.4.5
dill==0.3.5.1
Django==4.1
django-ckeditor==6.4.2
django-js-asset==2.0.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.2.0
platformdirs==2.5.2
pylint==2.14.5
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.2
typing-extensions==4.3.0
tzdata==2022.1
wrapt==1.14.1
